Move loading multiple environments to env.go
This way all the env related code is in the same place.
