+++ env.go
@@ -11,6 +11,33 @@ var envEntryRegexp = regexp.MustCompile("^([A-Za-z_0-9]+)=(.*)$")
 
 type Env map[string]string
 
+func loadEnvs(files []string) (Env, error) {
+	if len(files) == 0 {
+		env, err := ReadEnv(fullPath(".env"))
+		if err != nil {
+			return nil, err
+		} else {
+			return env, nil
+		}
+	}
+
+	// Handle multiple environment files
+	env := make(Env)
+	for _, file := range files {
+		tmpEnv, err := ReadEnv(file)
+
+		if err != nil {
+			return nil, err
+		}
+
+		// Merge the file I just read into the env.
+		for k, v := range tmpEnv {
+			env[k] = v
+		}
+	}
+	return env, nil
+}
+
 func ReadEnv(filename string) (Env, error) {
 	if _, err := os.Stat(filename); os.IsNotExist(err) {
 		return make(Env), nil
